Bump CMake min version for IBIze and clean up build files
Bump the CMake minimum version for IBIze to make it generate Makefiles
without errors or warnings about unsupported or deprecated CMake
versions. CMake 3.10 seems a reasonable minimum version here.

Delete the top-level CMake file, as IBIze is the only CMake project in
OpenJK-Tools. Add the project definition in the ibize subdirectory and
adjust the README accordingly.